arm_mpam: Add helper to reset saved mbwu state
resctrl expects to reset the bandwidth counters when the filesystem
is mounted.

To allow this, add a helper that clears the saved mbwu state. Instead
of cross calling to each CPU that can access the component MSC to
write to the counter, set a flag that causes it to be zero'd on the
the next read. This is easily done by forcing a configuration update.

Signed-off-by: James Morse <james.morse@arm.com>
(cherry picked from commit ca01e27727c44fc0deef570f3e5bc8e05e16e092 git://git.kernel.org/pub/scm/linux/kernel/git/morse/linux.git)
Signed-off-by: KobaK <kobak@nvidia.com>