FIXME arm_mpam: Make mpam_msmon_read() safe for irq-masked callers
The resctrl_pmu driver calls resctrl_arch_rmid_read() from interrupt
context. This eventually ends up in mpam_msmon_read() which cross-calls
to access remote cache-slices, and sleeps if the monitor returned
not-ready.

To support resctrl_pmu, remove these behaviours if irqs are masked.

Instead of cross-calling, return an error. This means the resctrl_pmu
can't be supported on platforms with cache-slices that aren't accessible
from every CPU. Instead of sleeping when the counter is busy, return
-EBUSY. The resctrl_pmu driver will not update the counter in this case.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: KobaK <kobak@nvidia.com>