fs/resctrl: Add cranky debug for reading CPU msr values
MPAM affects performance, which makes it hard to debug. (is that the
performance I was expecting?). One source of confusion is whether a
task is having its traffic labelled with the correct PARTID. This
is fairly straightforward for tasks, but once the resctrl 'cpus' files
get involved, the picture is more confusing.

As a debug aid, add a file in /proc that uses task_work to read the
CPU register for a task. This is output raw without interpretation.

This is not for upstream.

task_work may not be able to complete its work quickly if the task
is unable to take signals. Don't be too surprised if for some tasks
this appears to block the readers of the file.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Tushar Dave <tdave@nvidia.com>
(cherry picked from commit 1ac15de288a0fd61ab998fe4c6cd0ca90d09b056 https://github.com/NVIDIA-BaseOS-6/linux-nvidia-6.5/tree/MPAM-dev)
Signed-off-by: KobaK <kobak@nvidia.com>