fs/restrl: Allow the overflow handler to be disabled
Resctrl has an overflow handler that runs on each domain every second
to ensure that any overflow of the hardware counter is accounted for.

MPAM can have counters as large as 63 bits, in which case there is no
need to check for overflow.

Call the new arch helpers to determine this.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Shanker Donthineni <sdonthineni@nvidia.com>