fs/resctrl: Use schema format to check the resource is a bitmap
rdtgroup_cbm_to_size() uses a WARN_ON_ONCE() to assert that the resource
it has been passed is one of the L2 or L3 cache. This is to avoid using
uninitialised bitmap properties.

Updating this list for every resource that is configured by a bitmap
doesn't scale. Instead change the WARN_ON_ONCE() to use the schema
format the arch code requested for the resource.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Shanker Donthineni <sdonthineni@nvidia.com>