fs/resctrl: Move MBA supported check to parse_line() instead of parse_bw()
MBA is only supported on platforms where the delay inserted by the control
is linear. Resctrl checks the two properties provided by the arch code
match each time it parses part of a new control value.

This doesn't need to be done so frequently, and obscures changes to
parse_bw() to abstract it for use with other control types.

Move this check to the parse_line() caller so it only happens once.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Shanker Donthineni <sdonthineni@nvidia.com>