fs/resctrl: Add fflags_from_schema() for files based on schema format
MPAM has cache capacity controls that effectively take a percentage.
Resctrl supports percentages, but the collection of files that are
exposed to describe this control belong to the MB resource. New files
have been added that are selected based on the schema format.

Apply the flags to enable these files based on the schema format.
Add a new fflags_from_schema() that is used for controls.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Shanker Donthineni <sdonthineni@nvidia.com>