arm64: Update AMU-based freq scale factor on entering idle
Now that the frequency scale factor has been activated for retrieving
current frequency on a given CPU, trigger its update upon entering
idle. This will, to an extent, allow querying last known frequency
in a non-invasive way. It will also improve the frequency scale factor
accuracy when a CPU entering idle did not receive a tick for a while.
As a consequence, for idle cores, the reported frequency will be the
last one observed before entering the idle state.

Suggested-by: Vanshidhar Konda <vanshikonda@os.amperecomputing.com>
Signed-off-by: Beata Michalska <beata.michalska@arm.com>
Ref: https://patchwork.kernel.org/project/linux-arm-kernel/cover/20241206135600.4083965-1-beata.michalska@arm.com/
Signed-off-by: KobaK <kobak@nvidia.com>