fs/resctrl,x86/resctrl: Factor mba rounding to be per-arch
The control value parser for the MB resource currently coerces the
memory bandwidth percentage value from userspace to be an exact
multiple of the bw_gran parameter.

On MPAM systems, this results in somewhat worse-than-worst-case
rounding, since bw_gran is in general only an approximation to the
actual hardware granularity, and the hardware bandwidth allocation
control value is not natively a percentage.

Allow the arch to provide its own conversion that is appropriate for
the hardware, and move the existing conversion to x86.  This will avoid
accumulated error from rounding the value twice on MPAM systems.

Clarify the documentation, but avoid overly exact promises.

Clamping to bw_min and bw_max still feels generic: leave it in the core
code, for now.

No functional change.

Signed-off-by: Dave Martin <Dave.Martin@arm.com>
Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Shanker Donthineni <sdonthineni@nvidia.com>