arm_mpam: Add resctrl_arch_round_bw()
Add the required hook to pre-round a userspace memory bandwidth
allocation percentage value to a value acceptable to the driver backend.

For MPAM, no rounding is needed because the driver has all the
information necessary for rounding the value when
resctrl_arch_update_one() is called.

So, just "round" the value to itself here.

Signed-off-by: Dave Martin <Dave.Martin@arm.com>
Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Shanker Donthineni <sdonthineni@nvidia.com>