arm_mpam: resctrl: Pre-allocate assignable monitors
When there are not enough monitors, MPAM is able to emulate ABMC by making
a smaller number of monitors assignable. These monitors still need to be
allocated from the driver, and mapped to whichever control/monitor group
resctrl wants to use them with.

Add a second array to hold the monitor values indexed by resctrl's
cntr_id.

When CDP is in use, two monitors are needed so the available number of
counters halves. Platforms witih one monitor will have zero monitors
when CDP is in use.

Change-Id: I96df13dac7f8b64f458c14410eef504f17e00962
Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Shanker Donthineni <sdonthineni@nvidia.com>